+++ e2e/packages/client-vanilla/src/mud/setupNetwork.ts
@@ -4,8 +4,8 @@ import { encodeEntity, syncToRecs } from "@latticexyz/store-sync/recs";
 import { getNetworkConfig } from "./getNetworkConfig";
 import { world } from "./world";
 import { IWorld__factory } from "contracts/types/ethers-contracts/factories/IWorld__factory";
-import storeConfig from "contracts/mud.config";
 import { createBurnerAccount, createContract, transportObserver } from "@latticexyz/common";
+import mudConfig from "contracts/mud.config";
 
 export type SetupNetworkResult = Awaited<ReturnType<typeof setupNetwork>>;
 
@@ -17,8 +17,6 @@ export async function setupNetwork() {
     transport: transportObserver(http(networkConfig.rpcHttpUrl ?? undefined)),
     pollingInterval: 1000,
   } as const satisfies ClientConfig;
-
-  console.log("client options", clientOptions);
 
   const publicClient = createPublicClient(clientOptions);
 
@@ -37,7 +35,7 @@ export async function setupNetwork() {
 
   const { components, latestBlock$, blockStorageOperations$, waitForTransaction } = await syncToRecs({
     world,
-    config: storeConfig,
+    config: mudConfig,
     address: networkConfig.worldAddress as Hex,
     publicClient,
     startBlock: BigInt(networkConfig.initialBlockNumber),