+++ packages/block-logs-stream/src/groupLogsByBlockNumber.test.ts
@@ -1,53 +1,30 @@
 import { describe, it, expect } from "vitest";
 import { groupLogsByBlockNumber } from "./groupLogsByBlockNumber";
-import { Log } from "viem";
 
 describe("groupLogsByBlockNumber", () => {
   it("groups logs by block number and correctly sorts them", () => {
     const logs = [
       {
         blockNumber: 1n,
-        blockHash: "0x",
         logIndex: 4,
-        transactionHash: "0x",
-        transactionIndex: 0,
       },
       {
         blockNumber: 5n,
-        blockHash: "0x",
         logIndex: 0,
-        transactionHash: "0x",
-        transactionIndex: 0,
       },
       {
         blockNumber: 1n,
-        blockHash: "0x",
         logIndex: 0,
-        transactionHash: "0x",
-        transactionIndex: 0,
       },
       {
         blockNumber: 1n,
-        blockHash: "0x",
         logIndex: 2,
-        transactionHash: "0x",
-        transactionIndex: 0,
-      },
-      {
-        blockNumber: null,
-        blockHash: null,
-        logIndex: null,
-        transactionHash: null,
-        transactionIndex: null,
       },
       {
         blockNumber: 3n,
-        blockHash: "0x",
         logIndex: 3,
-        transactionHash: "0x",
-        transactionIndex: 0,
       },
-    ] as any as Log[];
+    ];
 
     expect(groupLogsByBlockNumber(logs)).toMatchInlineSnapshot(`
       [
@@ -55,25 +32,16 @@ describe("groupLogsByBlockNumber", () => {
           "blockNumber": 1n,
           "logs": [
             {
-              "blockHash": "0x",
               "blockNumber": 1n,
               "logIndex": 0,
-              "transactionHash": "0x",
-              "transactionIndex": 0,
             },
             {
-              "blockHash": "0x",
               "blockNumber": 1n,
               "logIndex": 2,
-              "transactionHash": "0x",
-              "transactionIndex": 0,
             },
             {
-              "blockHash": "0x",
               "blockNumber": 1n,
               "logIndex": 4,
-              "transactionHash": "0x",
-              "transactionIndex": 0,
             },
           ],
         },
@@ -81,11 +49,8 @@ describe("groupLogsByBlockNumber", () => {
           "blockNumber": 3n,
           "logs": [
             {
-              "blockHash": "0x",
               "blockNumber": 3n,
               "logIndex": 3,
-              "transactionHash": "0x",
-              "transactionIndex": 0,
             },
           ],
         },
@@ -93,11 +58,8 @@ describe("groupLogsByBlockNumber", () => {
           "blockNumber": 5n,
           "logs": [
             {
-              "blockHash": "0x",
               "blockNumber": 5n,
               "logIndex": 0,
-              "transactionHash": "0x",
-              "transactionIndex": 0,
             },
           ],
         },
@@ -109,12 +71,9 @@ describe("groupLogsByBlockNumber", () => {
     const logs = [
       {
         blockNumber: 1n,
-        blockHash: "0x",
         logIndex: 4,
-        transactionHash: "0x",
-        transactionIndex: 0,
       },
-    ] as any as Log[];
+    ];
 
     expect(groupLogsByBlockNumber(logs, 2n)).toMatchInlineSnapshot(`
       [
@@ -122,11 +81,8 @@ describe("groupLogsByBlockNumber", () => {
           "blockNumber": 1n,
           "logs": [
             {
-              "blockHash": "0x",
               "blockNumber": 1n,
               "logIndex": 4,
-              "transactionHash": "0x",
-              "transactionIndex": 0,
             },
           ],
         },
@@ -142,12 +98,9 @@ describe("groupLogsByBlockNumber", () => {
     const logs = [
       {
         blockNumber: 2n,
-        blockHash: "0x",
         logIndex: 4,
-        transactionHash: "0x",
-        transactionIndex: 0,
       },
-    ] as any as Log[];
+    ];
 
     expect(groupLogsByBlockNumber(logs, 2n)).toMatchInlineSnapshot(`
       [
@@ -155,11 +108,8 @@ describe("groupLogsByBlockNumber", () => {
           "blockNumber": 2n,
           "logs": [
             {
-              "blockHash": "0x",
               "blockNumber": 2n,
               "logIndex": 4,
-              "transactionHash": "0x",
-              "transactionIndex": 0,
             },
           ],
         },