+++ packages/dev-tools/src/recs/ComponentsPage.tsx
@@ -1,24 +1,28 @@
 import { Outlet, useNavigate, useParams } from "react-router-dom";
 import { NavButton } from "../NavButton";
-import { useTables } from "./useTables";
 import { useEffect, useRef } from "react";
 import { twMerge } from "tailwind-merge";
+import { useDevToolsContext } from "../DevToolsContext";
+import { getComponentName } from "./getComponentName";
 
-export function TablesPage() {
-  const detailsRef = useRef<HTMLDetailsElement>(null);
-  const tables = useTables();
-  const { table: tableParam } = useParams();
+export function ComponentsPage() {
+  const { recsWorld: world } = useDevToolsContext();
+  if (!world) throw new Error("Missing recsWorld");
+
+  const components = [...world.components].sort((a, b) => getComponentName(a).localeCompare(getComponentName(b)));
 
   // TODO: lift up selected component so we can remember previous selection between tab nav
-  const { tableId: selectedTableId, component: selectedComponent } =
-    tables.find(({ component }) => component === tableParam) ?? {};
+  const { id: idParam } = useParams();
+  const selectedComponent = components.find((component) => component.id === idParam) ?? components[0];
+
+  const detailsRef = useRef<HTMLDetailsElement>(null);
   const navigate = useNavigate();
 
   useEffect(() => {
-    if (tables.length && !selectedComponent) {
-      navigate(tables[0].component);
+    if (idParam !== selectedComponent.id) {
+      navigate(selectedComponent.id);
     }
-  }, [tables, selectedComponent]);
+  }, [idParam, selectedComponent.id]);
 
   useEffect(() => {
     const listener = (event: MouseEvent) => {
@@ -32,42 +36,40 @@ export function TablesPage() {
 
   return (
     <div className="p-6 space-y-4">
-      {!tables.length ? (
-        <>Waiting for tables…</>
+      {!components.length ? (
+        <>Waiting for components…</>
       ) : (
         <div className="space-y-2">
-          <h1 className="font-bold text-white/40 uppercase text-xs">Table</h1>
+          <h1 className="font-bold text-white/40 uppercase text-xs">Component</h1>
 
           <details ref={detailsRef} className="pointer-events-none select-none">
             <summary className="group pointer-events-auto cursor-pointer inline-flex">
               <span className="inline-flex gap-2 px-3 py-2 items-center border-2 border-white/10 rounded group-hover:border-blue-700 group-hover:bg-blue-700 group-hover:text-white">
-                {selectedTableId ? (
-                  <span className="font-mono">
-                    {selectedTableId.namespace}:{selectedTableId.name}
-                  </span>
+                {selectedComponent ? (
+                  <span className="font-mono">{getComponentName(selectedComponent)}</span>
                 ) : (
-                  <span>Pick a table…</span>
+                  <span>Pick a component…</span>
                 )}
                 <span className="text-white/40 text-xs">▼</span>
               </span>
             </summary>
             <div className="relative">
               <div className="pointer-events-auto absolute top-1 left-0 z-20 bg-slate-700 rounded shadow-lg flex flex-col py-1.5 font-mono text-xs leading-none">
-                {tables.map(({ component, tableId }) => (
+                {components.map((component) => (
                   <NavButton
                     className={twMerge(
                       "px-2 py-1.5 text-left hover:bg-blue-700 hover:text-white",
                       component === selectedComponent ? "bg-slate-600" : null
                     )}
-                    key={component}
-                    to={component}
+                    key={component.id}
+                    to={component.id}
                     onClick={() => {
                       if (detailsRef.current) {
                         detailsRef.current.open = false;
                       }
                     }}
                   >
-                    {tableId.namespace}:{tableId.name}
+                    {getComponentName(component)}
                   </NavButton>
                 ))}
               </div>