+++ packages/store-sync/src/recs/common.ts
@@ -4,16 +4,11 @@ import { SchemaAbiTypeToRecsType } from "./schemaAbiTypeToRecsType";
 import { SchemaAbiType } from "@latticexyz/schema-type";
 
 export type StoreComponentMetadata = {
+  componentName: string;
+  tableName: string;
   keySchema: KeySchema;
   valueSchema: ValueSchema;
 };
-
-export enum SyncStep {
-  INITIALIZE = "initialize",
-  SNAPSHOT = "snapshot",
-  RPC = "rpc",
-  LIVE = "live",
-}
 
 export type ConfigToRecsComponents<TConfig extends StoreConfig> = {
   [tableName in keyof TConfig["tables"] & string]: RecsComponent<